# Situationsverwaltung/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.2"
  title: SitOPT API Reference
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - text/html

# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
  /situations/occured:
    x-swagger-router-controller: situations
    put:
      tags:
        - situation
      summary: Change "occured"-attribute
      description: Change the "occured"-attribute of either specified or all situations. For testing purpose only!
      operationId: situationOccured
      
      parameters:
        - in: query
          name: ID
          description: ID of situation. If left empty, all situations change their "occured"-attribute.
          required: false
          type: string
       

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/byName:
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get situation by name
      description: Returns an array of situations with the specified name. Multiple results possible
      operationId: situationByName
      
      parameters:
        - in: query
          name: name
          description: Name of situation
          required: true
          type: string
       

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/ByThingAndTemplate:
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get situation by thing and situation template
      description: A situation is uniquely identifiable by the thing it belongs to and the situation template it derives from. Returns a document.
      operationId: situationByThingAndTemplate
      
      parameters:
        - name: thing
          in: query
          description: ID of thing
          required: true
          type: string
        - in: query
          name: situationtemplate
          description: ID of situation template
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/byID:
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get situation by ID
      description: Get specified situations
      operationId: situationByID
      
      parameters:
        - name: ID
          in: query
          description: ID of situation
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    delete:
      tags:
        - situation
      summary: Delete situation by ID
      description: Deletes specified situation 
      operationId: deleteSituationByID
      
      parameters:
        - name: ID
          in: query
          description: ID of situation
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/changes:
    x-swagger-router-controller: situations
    
    
    get:
      tags:
        - situation
      summary: Get registrations
      description: Specify CallbackURL to get only those registrations with the specified CallbackURL. If emtpy, returns all registrations.
      operationId: allRegistrations
      
      parameters:
        - name: CallbackURL
          in: query
          description: URL the changes are sent to.
          required: false
          type: string
          
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: string
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


    
    
    
    post:
      tags:
        - situation
      summary: Registrate for changes
      description: Get a notification when the situation - specified by corresponding thing & situationtemplate - changes via a callback to a specified URL. The parameter "continuous" controls whether the callback only happens once when the situation changes or continues forever. Either enter both IDs or none. 
      operationId: situationChange
      
      parameters:
        - name: SitTempID
          in: query
          description: ID of situation template. If empty, any situation change will be sent to the callback URL
          required: false
          type: string
          
        - name: ThingID
          in: query
          description: ID of thing. If empty, any situation change will be sent to the callback URL
          required: false
          type: string
          
        - name: CallbackURL
          in: query
          description: URL the callback is sent to
          required: true
          type: string
          
        - name: once
          in: query
          description: whether the callback continues or only once
          required: true
          type: boolean
          default: false

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "400":
          description: Bad Content
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    
    delete:        
      tags:
        - situation
      summary: Deletes registration
      description: Deletes the registration for either specified or all situation changes.
      
      operationId: situationChangeDelete
      
      parameters:
        - name: ID
          in: query
          description: ID of situation. If empty, callbacks to the specified URL will be stopped.
          required: false
          type: string
          
        - name: CallbackURL
          in: query
          description: URL to which the callback shall be stopped
          required: true
          type: string


      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations:
  # binds a127 app logic to a route
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get all situations
      description: Get all situations
      operationId: allSituations
      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/SituationModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
      
    post:
      tags:
        - situation
      summary: Store situation
      description: Stores situation in database. ID of thing, situation template and sensors referenced in the situation must be in the database.
      operationId: saveSituation
      # used as the method name of the controller
      parameters:
        - in: body
          name: body
          description: input
          required: false
          schema:
            $ref: "#/definitions/SituationModel"
            
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /things:
  # binds a127 app logic to a route
    x-swagger-router-controller: things
    
    get:
      tags:
        - thing
      summary: Get all things
      description: Get all things
      operationId: allThings
      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/ThingModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
            
    post:
      tags:
        - thing
      summary: Save thing
      description: Save things in the database. Field "location" uses GeoJSON. Only use value of GeoJSON field "geometry"
      operationId: saveThing
      # used as the method name of the controller
      parameters:
        - in: body
          name: body
          description: input
          required: false
          schema:
            $ref: "#/definitions/ThingModel"
            
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SituationModel:
    required:
      - thing
      - timestamp
      - situationtemplate
      - occured
      - sensorvalues
    properties:
      thing:
        type: string
      timestamp:
        type: string
      situationtemplate:
        type: string
      occured:
        type: boolean
      sensorvalues:
        type: array
        items:
          $ref: "#/definitions/SensorvalueModel"
  GetModel:
    required:
      - name
    properties:
      name:
        type: string
  ThingModel:
    required:
      - name
      - sensors
      - owners
    properties:
      state:
        type: array
        items:
          type: string
      name:
        type: string
      url:
        type: string
      description:
        type: string
      sensors:
        type: array
        items:
          type: string
      location:
        type: string
      owners:
        type: array
        items:
          type: string
...
  SensorvalueModel:
    required:
      - sensor
      - value
    properties:
      sensor:
        type: string 
      value:
        type: number
        format: float
      timestamp:
        type: string
      quality:
        type: number
        format: float
  Response:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string